requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity
azure-mgmt-compute==26.0.0
azure.mgmt.network
azure.mgmt.resource
dnspython
python-dotenv
